Updated  to include Remove-FsrmFileScreenException
Added a command to remove existing File Screen Exception before creating a new one.  If the contents of the ExcludePaths.txt changes this will correct the creation of the new screen.  For example, if a path was in the ExcludePaths.txt file and then is removed from it; the previous version of the script would not have removed that path from the Exception List and the location would still be excluded despite it no longer being in the ExcludePaths.txt file.  This corrects that oversight.